Split the admin listings template tests off the max-lines list
Break test/templates/admin/listings.test.ts (2,088 lines) into a
test/templates/admin/listings/ folder grouped by what each part renders,
and drop the file's entry from the noExcessiveLinesPerFile ignore list.

Shared render setup moves to helpers.ts (renderListingDetail, detailHtml,
editPanelHtml, the builder-env wrappers, TEST_SESSION and the beforeAll/
afterEach hooks). Bringing the file back under the duplication checker's
size limit surfaced repeated listing/attendee setup, now extracted into
small local fixtures so jscpd stays at zero.

Every original test carries over unchanged; the folder's 188 nodes pass,
the linter and duplication checker are clean.
