+++ biome.json
@@ -178,8 +178,7 @@
         "test/lib/server-payments.test.ts",
         "test/lib/square.test.ts",
         "test/lib/stripe.test.ts",
-        "test/routes/api.test.ts",
-        "test/templates/admin/listings.test.ts"
+        "test/routes/api.test.ts"
       ],
       "linter": {
         "rules": {
