--- test/templates/admin/listings.test.ts
@@ -1,2088 +1,0 @@
-import { expect } from "@std/expect";
-import { afterEach, beforeAll, describe, it as test } from "@std/testing/bdd";
-import { signCsrfToken } from "#shared/csrf.ts";
-import { attendeeNameMap } from "#shared/db/system-notes.ts";
-import { detectIframeMode } from "#shared/iframe.ts";
-import { account } from "#shared/ledger/account.ts";
-import { runWithStorageConfig } from "#shared/storage.ts";
-import { emptyLedgerNames } from "#templates/admin/ledger.tsx";
-import {
-  adminListingRecalculatePage,
-  nearCapacity,
-} from "#templates/admin/listings/aggregates.tsx";
-import {
-  completePaymentAttendees,
-  isIncompletePayment,
-} from "#templates/admin/listings/attendees.tsx";
-import { ListingEditPanel } from "#templates/admin/listings/edit-panel.tsx";
-import {
-  adminDuplicateListingPage,
-  adminListingNewPage,
-} from "#templates/admin/listings/form-pages.tsx";
-import {
-  ListingDeactivatedBanner,
-  ListingOverviewPanel,
-  overviewStatsFromAttendees,
-  overviewStatsFromDbStats,
-} from "#templates/admin/listings/overview.tsx";
-import { ListingRosterPanel } from "#templates/admin/listings/roster.tsx";
-import { getListingFields } from "#templates/fields.ts";
-import {
-  describeWithEnv,
-  setTestEnv,
-  setupTestEncryptionKey,
-  TEST_STORAGE_ZONE,
-  testAttendee,
-  testGroup,
-  testListingWithCount,
-  withStorageDisabled,
-} from "#test-utils";
-
-const TEST_SESSION = { adminLevel: "owner" as const };
-
-/** Run fn with CAN_BUILD_SITES pinned to `value` in this worker's env overlay
- *  — never the real process env, which every parallel test worker shares. */
-const withBuilderEnv =
-  (value: string | undefined) =>
-  (fn: () => void): void => {
-    const restore = setTestEnv({ CAN_BUILD_SITES: value });
-    try {
-      fn();
-    } finally {
-      restore();
-    }
-  };
-const withBuilder = withBuilderEnv("true");
-const withoutBuilder = withBuilderEnv(undefined);
-
-/** Render the listing detail view the way the entity page composes it: the
- *  Overview panel (details table, income/ledger) plus the Roster panel (attendee
- *  table, filters, failed payments, add-attendee). The legacy `adminListingPage`
- *  composer was removed, so these tests assert against the live panels. */
-const renderListingDetail = (
-  opts: Parameters<typeof ListingRosterPanel>[0],
-): string =>
-  String(
-    ListingOverviewPanel({
-      aggregateRecalculation: opts.aggregateRecalculation,
-      allowedDomain: opts.allowedDomain,
-      groupContext: opts.groupContext,
-      isChild: opts.isChild,
-      isHiddenPackageMember: opts.isHiddenPackageMember,
-      ledger: opts.ledger,
-      listing: opts.listing,
-      // The Overview now takes precomputed stats + note-author names instead of
-      // the raw attendee list; derive them from the fixture's attendees so these
-      // tests exercise the same rendered output the SQL path produces.
-      noteNames: attendeeNameMap(opts.attendees),
-      questionData: opts.questionData,
-      revenueBreakdown: opts.revenueBreakdown,
-      stats: overviewStatsFromAttendees(opts.listing, opts.attendees),
-      systemNotes: opts.systemNotes,
-    }),
-  ) + String(ListingRosterPanel(opts));
-
-beforeAll(async () => {
-  setupTestEncryptionKey();
-  await signCsrfToken();
-});
-
-afterEach(() => {
-  detectIframeMode("https://example.com/");
-});
-
-describe("adminListingEditPage group select", () => {
-  test("checks the listing's groups when groups exist", () => {
-    const groups = [testGroup({ id: 2, name: "Group Two" })];
-    const listing = testListingWithCount({});
-    const html = String(
-      ListingEditPanel({
-        groups,
-        listing,
-        selectedGroupIds: [2],
-        session: TEST_SESSION,
-      }),
-    );
-    expect(html).toContain('name="group_ids"');
-    expect(html).toContain('value="2"');
-    expect(html).toContain("checked");
-  });
-
-  test("does not link to the JSON export (that now lives on the Actions tab)", () => {
-    const listing = testListingWithCount({ id: 9 });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    // The Actions tab is now editor-visible too, so the export link lives
-    // only there — the Edit panel no longer duplicates it.
-    expect(html).not.toContain(`/admin/listing/${listing.id}/export.json`);
-  });
-});
-
-describe("adminListingEditPage duration warning", () => {
-  test("includes duration warning + confirmation gate with current duration", () => {
-    const listing = testListingWithCount({
-      duration_days: 3,
-      listing_type: "daily",
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain(
-      "Changing booking duration will update existing bookings",
-    );
-    expect(html).toContain('id="duration-warning"');
-    expect(html).toContain('id="duration-warning-confirm"');
-    // The current duration is exposed via a data attribute so the bundled
-    // admin script can compare against the form's input.
-    expect(html).toContain('data-duration-original="3"');
-  });
-});
-
-describe("adminListingEditPage day prices", () => {
-  test("renders priced day-count inputs and checks the customisable toggle", () => {
-    const listing = testListingWithCount({
-      customisable_days: true,
-      day_prices: { 1: 1000, 2: 1800 },
-      duration_days: 2,
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    // One input per day up to the maximum duration, pre-filled from day_prices.
-    expect(html).toContain('name="day_price_1"');
-    expect(html).toContain('value="10.00"');
-    expect(html).toContain('name="day_price_2"');
-    expect(html).toContain('value="18.00"');
-    // The customisable-days checkbox is rendered checked for such a listing.
-    expect(html).toContain('name="customisable_days"');
-    expect(html).toContain("Day Prices (customisable days)");
-  });
-
-  test("renders a single blank day-price row on the new-listing form", () => {
-    const html = adminListingNewPage([], TEST_SESSION);
-    expect(html).toContain('name="day_price_1"');
-    // The maximum defaults to 1 day for a new listing, so only one row shows.
-    expect(html).not.toContain('name="day_price_2"');
-  });
-});
-
-describe("adminListingEditPage form sections", () => {
-  test("groups fields under section legends and an Advanced disclosure", () => {
-    const listing = testListingWithCount();
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain("<legend>Basics</legend>");
-    expect(html).toContain("<legend>Tickets &amp; Pricing</legend>");
-    expect(html).toContain("<legend>Daily Scheduling</legend>");
-    expect(html).toContain(
-      "<legend>Booking Duration &amp; Day Prices</legend>",
-    );
-    expect(html).toContain("<legend>Options &amp; Visibility</legend>");
-    expect(html).toContain("<summary>Advanced settings</summary>");
-  });
-
-  test("renders the day-prices block immediately after the customisable-days checkbox", () => {
-    const listing = testListingWithCount({
-      customisable_days: true,
-      day_prices: { 1: 1000 },
-      duration_days: 1,
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    const customisableIdx = html.indexOf('name="customisable_days"');
-    const dayPriceIdx = html.indexOf('name="day_price_1"');
-    const contactFieldsIdx = html.indexOf('name="fields"');
-    // The prices follow the checkbox, and both sit before the later Options
-    // section — i.e. no longer dumped at the bottom of the form.
-    expect(customisableIdx).toBeGreaterThan(-1);
-    expect(customisableIdx).toBeLessThan(dayPriceIdx);
-    expect(dayPriceIdx).toBeLessThan(contactFieldsIdx);
-  });
-
-  test("places the technical fields inside the Advanced disclosure", () => {
-    const listing = testListingWithCount();
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    const advancedIdx = html.indexOf("<summary>Advanced settings</summary>");
-    expect(advancedIdx).toBeGreaterThan(-1);
-    expect(advancedIdx).toBeLessThan(html.indexOf('name="webhook_url"'));
-    expect(advancedIdx).toBeLessThan(html.indexOf('name="thank_you_url"'));
-    expect(advancedIdx).toBeLessThan(html.indexOf('name="slug"'));
-  });
-
-  test("renders editable running totals with a recalculation link", () => {
-    const listing = testListingWithCount({
-      attendee_count: 7,
-      tickets_count: 3,
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain("<legend>Running totals</legend>");
-    expect(html).toContain("Accuracy is not guaranteed");
-    expect(html).toContain('name="booked_quantity"');
-    expect(html).toContain('value="7"');
-    expect(html).toContain('name="tickets_count"');
-    expect(html).toContain('value="3"');
-    // Income is not a count override in the running-totals form — that form keeps
-    // only the two count aggregates, which post to the recalculate route.
-    const totalsForm = html.slice(
-      html.indexOf("<legend>Running totals</legend>"),
-      html.indexOf("Adjust income"),
-    );
-    expect(totalsForm).not.toContain('name="income"');
-    expect(html).toContain(`/admin/listings/recalculate/${listing.id}`);
-  });
-
-  test("renders the separate income-correction form (decision 14)", () => {
-    const listing = testListingWithCount();
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    // Income correction is restored as a dedicated warned form that posts a
-    // writeoff adjustment to the money ledger, kept apart from the counts override.
-    expect(html).toContain("<h2>Adjust income</h2>");
-    expect(html).toContain(`action="/admin/listing/${listing.id}/income"`);
-    expect(html).toContain('name="income"');
-    expect(html).toContain("correcting entry to the money ledger");
-  });
-
-  test("links from the income form to the detail page's income & ledger breakdown", () => {
-    const listing = testListingWithCount({ id: 4 });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    // A compact pointer beside the adjust-income form to the full reconciliation
-    // section on the detail page, so the two figures are explained in one place.
-    expect(html).toContain('href="/admin/listing/4#income-ledger"');
-    expect(html).toContain("Income &amp; ledger breakdown");
-  });
-
-  test("shows a running-total mismatch on the edit page", () => {
-    const listing = testListingWithCount({
-      attendee_count: 7,
-      tickets_count: 3,
-    });
-    const html = String(
-      ListingEditPanel({
-        aggregateRecalculation: {
-          booked_quantity: { current: 7, recalculated: 4 },
-          tickets_count: { current: 3, recalculated: 3 },
-        },
-        groups: [],
-        listing,
-        session: TEST_SESSION,
-      }),
-    );
-    expect(html).toContain("Mismatch");
-    expect(html).toContain("expected <strong>4</strong>, got");
-    expect(html).toContain(`/admin/listings/recalculate/${listing.id}`);
-  });
-});
-
-describe("adminListingRecalculatePage", () => {
-  test("shows current and attendee-derived totals with checkboxes", () => {
-    const listing = testListingWithCount({ name: "Workshop" });
-    const html = adminListingRecalculatePage(
-      listing,
-      {
-        booked_quantity: { current: 9, recalculated: 4 },
-        tickets_count: { current: 5, recalculated: 2 },
-      },
-      TEST_SESSION,
-    );
-    expect(html).toContain("Recalculate: Workshop");
-    expect(html).toContain("Current");
-    expect(html).toContain("From attendee data");
-    expect(html).toContain("Compare the stored listing totals");
-    expect(html).toContain('class="table-scroll"');
-    expect(html).toContain('name="recalculate_fields"');
-    expect(html).toContain('value="booked_quantity"');
-    expect(html).toContain(">9<");
-    expect(html).toContain(">4<");
-  });
-});
-
-describe("adminListingEditPage Advanced section auto-open", () => {
-  test("stays collapsed when only the slug is set", () => {
-    // Slug is always populated; on its own it must not force the section open.
-    const listing = testListingWithCount({
-      thank_you_url: "",
-      webhook_url: "",
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain('<details class="listing-advanced">');
-    expect(html).not.toContain('<details class="listing-advanced" open>');
-  });
-
-  test("opens when a thank-you URL is set", () => {
-    const listing = testListingWithCount({
-      thank_you_url: "https://example.com/thanks",
-      webhook_url: "",
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain('<details class="listing-advanced" open>');
-  });
-
-  test("opens when a webhook URL is set", () => {
-    const listing = testListingWithCount({
-      thank_you_url: "",
-      webhook_url: "https://hooks.example.com/notify",
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain('<details class="listing-advanced" open>');
-  });
-
-  test("opens on a validation error so hidden fields stay reachable", () => {
-    const listing = testListingWithCount({
-      thank_you_url: "",
-      webhook_url: "",
-    });
-    const html = String(
-      ListingEditPanel({
-        error: "Bad input",
-        groups: [],
-        listing,
-        session: TEST_SESSION,
-      }),
-    );
-    expect(html).toContain('<details class="listing-advanced" open>');
-  });
-});
-
-describe("adminListingNewPage Advanced section", () => {
-  test("renders collapsed by default", () => {
-    const html = adminListingNewPage([], TEST_SESSION);
-    expect(html).toContain('<details class="listing-advanced">');
-    expect(html).not.toContain('<details class="listing-advanced" open>');
-  });
-
-  test("opens when re-rendered with an error", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      error: "Something went wrong",
-    });
-    expect(html).toContain('<details class="listing-advanced" open>');
-  });
-});
-
-describe("adminListingEditPage Advanced section auto-open (builder fields)", () => {
-  test("opens when a renewal tier (months per unit) is set", () => {
-    withBuilder(() => {
-      const listing = testListingWithCount({
-        months_per_unit: 3,
-        thank_you_url: "",
-        webhook_url: "",
-      });
-      const html = String(
-        ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-      );
-      expect(html).toContain('<details class="listing-advanced" open>');
-    });
-  });
-
-  test("opens when initial site months is set", () => {
-    withBuilder(() => {
-      const listing = testListingWithCount({
-        initial_site_months: 6,
-        months_per_unit: 0,
-        thank_you_url: "",
-        webhook_url: "",
-      });
-      const html = String(
-        ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-      );
-      expect(html).toContain('<details class="listing-advanced" open>');
-    });
-  });
-
-  test("opens when a built site is assigned on booking", () => {
-    withBuilder(() => {
-      const listing = testListingWithCount({
-        assign_built_site: true,
-        initial_site_months: 0,
-        months_per_unit: 0,
-        thank_you_url: "",
-        webhook_url: "",
-      });
-      const html = String(
-        ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-      );
-      expect(html).toContain('<details class="listing-advanced" open>');
-    });
-  });
-
-  test("stays collapsed when no advanced field is set even with the builder enabled", () => {
-    withBuilder(() => {
-      const listing = testListingWithCount({
-        assign_built_site: false,
-        initial_site_months: 0,
-        months_per_unit: 0,
-        thank_you_url: "",
-        webhook_url: "",
-      });
-      const html = String(
-        ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-      );
-      expect(html).toContain('<details class="listing-advanced">');
-      expect(html).not.toContain('<details class="listing-advanced" open>');
-    });
-  });
-});
-
-describe("adminListingPage duration display", () => {
-  test("shows booking duration row for daily listings", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      duration_days: 3,
-      listing_type: "daily",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Booking Duration");
-    expect(html).toContain("3 day(s)");
-  });
-
-  test("omits booking duration row for standard listings", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      duration_days: 1,
-      listing_type: "standard",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("Booking Duration");
-  });
-
-  test("shows the customisable-days prices on a customisable listing", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      customisable_days: true,
-      day_prices: { 1: 1000, 2: 1800 },
-      duration_days: 2,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Customisable Days");
-    expect(html).toContain("1 day:");
-    expect(html).toContain("2 days:");
-  });
-
-  test("offers a day-count selector when adding to a customisable daily listing", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      customisable_days: true,
-      day_prices: { 1: 1000, 2: 1800 },
-      duration_days: 2,
-      listing_type: "daily",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain('name="day_count"');
-    expect(html).toContain("Number of days");
-  });
-
-  test("notes when a customisable listing has no day prices set", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      customisable_days: true,
-      day_prices: {},
-      duration_days: 3,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("No day prices set");
-  });
-
-  test("omits the customisable-days row for a fixed-duration listing", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      customisable_days: false,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("Customisable Days");
-  });
-});
-
-describe("adminListingPage", () => {
-  const listing = testListingWithCount({ attendee_count: 2 });
-
-  test("renders listing name", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Test Listing");
-  });
-
-  test("shows the listing ticket price in the details table", () => {
-    const paidListing = testListingWithCount({ unit_price: 1250 });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing: paidListing,
-    });
-    expect(html).toContain("Ticket Price");
-    expect(html).toContain("£12.50");
-  });
-
-  test("shows pay-more price bounds in the details table", () => {
-    const payMoreListing = testListingWithCount({
-      can_pay_more: true,
-      max_price: 2500,
-      unit_price: 1000,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing: payMoreListing,
-    });
-    expect(html).toContain("Ticket Price");
-    expect(html).toContain("£10");
-    expect(html).toContain("pay more: £10–£25");
-  });
-
-  test("shows pay-more enabled when no higher maximum is configured", () => {
-    const payMoreListing = testListingWithCount({
-      can_pay_more: true,
-      max_price: 0,
-      unit_price: 0,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing: payMoreListing,
-    });
-    expect(html).toContain("Ticket Price");
-    expect(html).toContain("Free");
-    expect(html).toContain("pay more enabled");
-  });
-
-  test("shows free when the listing has no ticket price", () => {
-    const freeListing = testListingWithCount({ unit_price: 0 });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing: freeListing,
-    });
-    expect(html).toContain("Ticket Price");
-    expect(html).toContain("Free");
-  });
-
-  test("shows attendees row with count and remaining", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Listing Attendees");
-    expect(html).toContain("2 / 100");
-    expect(html).toContain("98 remain");
-  });
-
-  test("shows a running-total mismatch in the details table", () => {
-    const html = renderListingDetail({
-      aggregateRecalculation: {
-        booked_quantity: { current: 2, recalculated: 1 },
-        tickets_count: { current: 0, recalculated: 0 },
-      },
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Running total check");
-    expect(html).toContain("expected <strong>1</strong>, got");
-    expect(html).toContain("Mismatch");
-    expect(html).not.toContain("Click for info");
-  });
-
-  test("renders no Group Attendees row when groupContext is omitted", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("Group Attendees");
-  });
-
-  test("shows Group Attendees row with count, cap, remaining, and link", () => {
-    const group = testGroup({
-      id: 7,
-      max_attendees: 50,
-      name: "Summer Festival",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      groupContext: { attendeeCount: 30, group },
-      listing,
-    });
-    expect(html).toContain("Group Attendees");
-    expect(html).toContain("30 / 50");
-    expect(html).toContain("20 remain");
-    expect(html).toContain('href="/admin/groups/7"');
-    expect(html).toContain("Summer Festival");
-    expect(html).toContain("across all listings");
-  });
-
-  test("Group Attendees row gets danger-text when at or near cap", () => {
-    const group = testGroup({ id: 8, max_attendees: 10 });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      groupContext: { attendeeCount: 10, group },
-      listing,
-    });
-    expect(html).toContain("danger-text");
-    expect(html).toContain("10 / 10");
-    expect(html).toContain("0 remain");
-  });
-
-  test("shows checked in row with 0 of 0 when no attendees", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Checked In");
-    expect(html).toContain("0 / 0");
-    expect(html).toContain("0 remain");
-  });
-
-  test("shows checked in count and remaining", () => {
-    const attendees = [
-      testAttendee({ checked_in: true, id: 1 }),
-      testAttendee({ checked_in: false, id: 2 }),
-      testAttendee({ checked_in: false, id: 3 }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("Checked In");
-    expect(html).toContain("1 / 3");
-    expect(html).toContain("2 remain");
-  });
-
-  test("roster keeps one row per booking line, with no Listings column", () => {
-    const attendees = [testAttendee({ id: 8, listing_id: listing.id })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    // The roster is scoped to one listing, so the Listings column stays off
-    // and each line keeps its own per-listing check-in action.
-    expect(html).not.toContain("<th>Listings</th>");
-    expect(html).toContain(`/admin/listing/${listing.id}/attendee/8/checkin`);
-  });
-
-  test("shows dual checked-in rows when attendees have multi-quantity", () => {
-    const attendees = [
-      testAttendee({ checked_in: true, id: 1, quantity: 2 }),
-      testAttendee({ checked_in: false, id: 2, quantity: 3 }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    // Tickets Checked In: 1 row / 2 rows, 1 remain
-    expect(html).toContain("Tickets Checked In");
-    expect(html).toContain("1 / 2");
-    expect(html).toContain("1 remain");
-    // Attendees Checked In: 2 qty / 5 total qty, 3 remain
-    expect(html).toContain("Attendees Checked In");
-    expect(html).toContain("2 / 5");
-    expect(html).toContain("3 remain");
-  });
-
-  test("dual checked-in rows show daily suffix when daily listing with dateFilter", () => {
-    const dailyListing = testListingWithCount({
-      attendee_count: 5,
-      listing_type: "daily",
-    });
-    const attendees = [
-      testAttendee({ checked_in: true, id: 1, quantity: 2 }),
-      testAttendee({ checked_in: false, id: 2, quantity: 3 }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      dateFilter: "2026-03-15",
-      listing: dailyListing,
-    });
-    expect(html).toContain("Attendees Checked In (Sunday 15 March 2026)");
-    expect(html).toContain("Tickets Checked In (Sunday 15 March 2026)");
-  });
-
-  test("dual checked-in rows show total suffix when daily listing without dateFilter", () => {
-    const dailyListing = testListingWithCount({
-      attendee_count: 5,
-      listing_type: "daily",
-    });
-    const attendees = [
-      testAttendee({ checked_in: true, id: 1, quantity: 2 }),
-      testAttendee({ checked_in: false, id: 2, quantity: 3 }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing: dailyListing,
-    });
-    expect(html).toContain("Attendees Checked In (total)");
-    expect(html).toContain("Tickets Checked In (total)");
-  });
-
-  test("shows thank you URL in copyable input", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Thank You URL");
-    expect(html).toContain('value="https://example.com/thanks"');
-    expect(html).toContain("readonly");
-    expect(html).toContain("data-select-on-click");
-  });
-
-  test("shows public URL with allowed domain", () => {
-    const html = renderListingDetail({
-      allowedDomain: "example.com",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Public URL");
-    expect(html).toContain('href="https://example.com/ticket/ab12c"');
-    expect(html).toContain("example.com/ticket/ab12c");
-  });
-
-  test("shows collapsed embed codes with allowed domain and iframe param", () => {
-    const html = renderListingDetail({
-      allowedDomain: "example.com",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain('for="embed-toggle-1"');
-    expect(html).toContain('class="embed-toggle-badge"');
-    expect(html).toContain('class="visually-hidden listing-embed-toggle"');
-    expect(html).toContain('class="listing-embed-row"');
-    expect(html).toContain("Embed Script");
-    expect(html).toContain("Embed Iframe");
-    expect(html).toContain("embed.js");
-    expect(html).toContain("data-listings=");
-    expect(html).toContain("https://example.com/ticket/ab12c?iframe=true");
-    expect(html).toContain("height: 600px");
-    expect(html).toContain("loading=");
-    expect(html).toContain("readonly");
-  });
-
-  test("iframe embed is a plain iframe without resizer scripts", () => {
-    const html = renderListingDetail({
-      allowedDomain: "example.com",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("iframeResize");
-  });
-
-  test("renders empty attendees state", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("No attendees yet");
-  });
-
-  test("renders attendees table", () => {
-    const attendees = [testAttendee()];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("John Doe");
-    expect(html).toContain("john@example.com");
-  });
-
-  test("escapes attendee data", () => {
-    const attendees = [testAttendee({ name: "<script>evil()</script>" })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("&lt;script&gt;");
-  });
-
-  test("includes back link", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("/admin/");
-  });
-
-  test("shows phone column when attendee has phone", () => {
-    const attendees = [testAttendee({ phone: "+1 555 123 4567" })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("<th>Phone</th>");
-    expect(html).toContain("+1 555 123 4567");
-  });
-
-  test("hides phone column when no attendees have phone", () => {
-    const attendees = [testAttendee({ phone: "" })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).not.toContain("<th>Phone</th>");
-  });
-
-  test("hides email column when no attendees have email", () => {
-    const attendees = [testAttendee({ email: "" })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("John Doe");
-    expect(html).not.toContain("<th>Email</th>");
-  });
-
-  test("shows danger-text class when near capacity", () => {
-    const nearFullListing = testListingWithCount({
-      attendee_count: 91,
-      max_attendees: 100,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing: nearFullListing,
-    });
-    expect(html).toContain('class="danger-text"');
-    expect(html).toContain("9 remain");
-  });
-
-  test("does not show danger-text class when not near capacity", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain('class="danger-text"');
-  });
-
-  test("shows deactivated alert for inactive listings", () => {
-    const html = String(ListingDeactivatedBanner({ active: false }));
-    expect(html).toContain('class="error"');
-    expect(html).toContain("This listing is deactivated and cannot be booked");
-  });
-
-  test("does not show deactivated alert for active listings", () => {
-    const html = String(ListingDeactivatedBanner({ active: true }));
-    expect(html).not.toContain(
-      "This listing is deactivated and cannot be booked",
-    );
-  });
-
-  test("shows ticket column header", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("<th>Ticket</th>");
-  });
-
-  test("shows ticket token as link to public ticket URL", () => {
-    const attendees = [testAttendee({ ticket_token: "abc123" })];
-    const html = renderListingDetail({
-      allowedDomain: "mysite.com",
-      attendees,
-      listing,
-    });
-    expect(html).toContain('href="https://mysite.com/t/abc123"');
-    expect(html).toContain("abc123");
-  });
-
-  test("renders empty date cell for attendee without date on daily listing", () => {
-    const dailyListing = testListingWithCount({
-      attendee_count: 1,
-      listing_type: "daily",
-    });
-    const attendees = [testAttendee({ date: null })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing: dailyListing,
-    });
-    expect(html).toContain("<th>Date</th>");
-  });
-
-  test("shows unlimited booking window when maximum_days_after is 0", () => {
-    const dailyListing = testListingWithCount({
-      listing_type: "daily",
-      maximum_days_after: 0,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing: dailyListing,
-    });
-    expect(html).toContain("unlimited");
-  });
-
-  test("shows numeric booking window when maximum_days_after is nonzero", () => {
-    const dailyListing = testListingWithCount({
-      listing_type: "daily",
-      maximum_days_after: 30,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing: dailyListing,
-    });
-    expect(html).toContain("to 30 days");
-    expect(html).not.toContain("unlimited");
-  });
-
-  test("shows danger-text for daily listing at capacity with date filter", () => {
-    const dailyListing = testListingWithCount({
-      attendee_count: 0,
-      listing_type: "daily",
-      max_attendees: 2,
-    });
-    const attendees = [testAttendee(), testAttendee({ id: 2, name: "Jane" })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      dateFilter: "2026-03-15",
-      listing: dailyListing,
-    });
-    expect(html).toContain('class="danger-text"');
-    expect(html).toContain("0 remain");
-  });
-});
-
-describe("adminListingNewPage", () => {
-  test("renders create listing form fields", () => {
-    const html = adminListingNewPage([], TEST_SESSION);
-    expect(html).toContain("Add Listing");
-    expect(html).toContain('name="name"');
-    expect(html).toContain('name="max_attendees"');
-    expect(html).toContain('name="thank_you_url"');
-    expect(html).toContain('name="unit_price"');
-    expect(html).toContain("Ticket Price");
-  });
-
-  test("renders breadcrumb back link", () => {
-    const html = adminListingNewPage([], TEST_SESSION);
-    expect(html).toContain('href="/admin/"');
-    expect(html).toContain("Listings");
-  });
-
-  test("renders group checkboxes when groups exist", () => {
-    const groups = [testGroup({ id: 2, name: "My Group" })];
-    const html = adminListingNewPage(groups, TEST_SESSION);
-    expect(html).toContain('name="group_ids"');
-    expect(html).toContain('value="2"');
-    expect(html).toContain("My Group");
-  });
-
-  test("renders error when provided", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      error: "Something went wrong",
-    });
-    expect(html).toContain("Something went wrong");
-  });
-
-  test("applies listing-form--hide-type class for templates with a fixed listing_type", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      templateId: "weekly-event",
-    });
-    expect(html).toContain("listing-form--hide-type");
-  });
-
-  test("does not apply listing-form--hide-type for templates with no fixed listing_type", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      templateId: "hireable-item",
-    });
-    expect(html).not.toContain("listing-form--hide-type");
-    expect(html).toContain("listing-form--templated");
-  });
-
-  test("seeds the hireable-item contact fields with phone for delivery contact", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      templateId: "hireable-item",
-    });
-    expect(html).toContain('name="fields" value="email" checked');
-    expect(html).toContain('name="fields" value="phone" checked');
-    expect(html).toContain('name="fields" value="address" checked');
-  });
-
-  test("seeds every weekday into the hireable-item bookable_days", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      templateId: "hireable-item",
-    });
-    for (const day of [
-      "Monday",
-      "Tuesday",
-      "Wednesday",
-      "Thursday",
-      "Friday",
-      "Saturday",
-      "Sunday",
-    ]) {
-      expect(html).toContain(`name="bookable_days" value="${day}" checked`);
-    }
-  });
-
-  test("applies listing-form--no-daily class for non-daily templates", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      templateId: "one-off-event",
-    });
-    expect(html).toContain("listing-form--no-daily");
-  });
-
-  test("does not apply listing-form--no-daily for templates without a fixed non-daily type", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      templateId: "hireable-item",
-    });
-    expect(html).not.toContain("listing-form--no-daily");
-  });
-
-  test("preserves selected groups on error re-render", () => {
-    const groups = [testGroup({ id: 3, name: "Group Three" })];
-    const html = adminListingNewPage(groups, TEST_SESSION, {
-      selectedGroupIds: [3],
-    });
-    expect(html).toContain('value="3"');
-    expect(html).toContain("checked");
-  });
-
-  test("carries custom sentinel through as template_id hidden input", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      templateId: "custom",
-    });
-    expect(html).toContain('name="template_id"');
-    expect(html).toContain('value="custom"');
-  });
-
-  test("carries duplicated_from hidden input when present in submitted values", () => {
-    const html = adminListingNewPage([], TEST_SESSION, {
-      values: { duplicated_from: "42" },
-    });
-    expect(html).toContain('name="duplicated_from"');
-    expect(html).toContain('value="42"');
-  });
-
-  test("does not render duplicated_from input when not in submitted values", () => {
-    const html = adminListingNewPage([], TEST_SESSION, { values: {} });
-    expect(html).not.toContain('name="duplicated_from"');
-  });
-});
-
-describe("adminListingPage export button", () => {
-  test("renders export CSV button", () => {
-    const listing = testListingWithCount({ attendee_count: 2 });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("/admin/listing/1/export");
-    expect(html).toContain("Export CSV");
-  });
-
-  test("the export link carries the active check-in filter", () => {
-    const html = renderListingDetail({
-      activeFilter: "in",
-      allowedDomain: "localhost",
-      attendees: [testAttendee({ checked_in: true })],
-      listing: testListingWithCount({ attendee_count: 1 }),
-    });
-    expect(html).toContain("/admin/listing/1/export?checkin=in");
-  });
-});
-
-describe("adminListingPage filter links", () => {
-  test("renders All / Checked In / Checked Out links", () => {
-    const listing = testListingWithCount({ attendee_count: 1 });
-    const attendees = [testAttendee()];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("All");
-    expect(html).toContain("Checked In");
-    expect(html).toContain("Checked Out");
-  });
-
-  test("bolds All when no filter is active", () => {
-    const listing = testListingWithCount({ attendee_count: 1 });
-    const attendees = [testAttendee()];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("<strong>All</strong>");
-    expect(html).toContain(
-      `href="/admin/listing/${listing.id}/attendees?filter=in"`,
-    );
-    expect(html).toContain(
-      `href="/admin/listing/${listing.id}/attendees?filter=out"`,
-    );
-  });
-
-  test("bolds Checked In when filter is in", () => {
-    const listing = testListingWithCount({ attendee_count: 1 });
-    const attendees = [testAttendee()];
-    const html = renderListingDetail({
-      activeFilter: "in",
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("<strong>Checked In</strong>");
-    expect(html).toContain(`href="/admin/listing/${listing.id}/attendees"`);
-  });
-
-  test("bolds Checked Out when filter is out", () => {
-    const listing = testListingWithCount({ attendee_count: 1 });
-    const attendees = [testAttendee()];
-    const html = renderListingDetail({
-      activeFilter: "out",
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("<strong>Checked Out</strong>");
-  });
-
-  test("filters to only checked-in attendees when filter is in", () => {
-    const listing = testListingWithCount({ attendee_count: 2 });
-    const attendees = [
-      testAttendee({ checked_in: true, id: 1, name: "Checked In User" }),
-      testAttendee({ checked_in: false, id: 2, name: "Not Checked In User" }),
-    ];
-    const html = renderListingDetail({
-      activeFilter: "in",
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("Checked In User");
-    expect(html).not.toContain("Not Checked In User");
-  });
-
-  test("filters to only checked-out attendees when filter is out", () => {
-    const listing = testListingWithCount({ attendee_count: 2 });
-    const attendees = [
-      testAttendee({ checked_in: true, id: 1, name: "Alice InPerson" }),
-      testAttendee({ checked_in: false, id: 2, name: "Bob Remote" }),
-    ];
-    const html = renderListingDetail({
-      activeFilter: "out",
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).not.toContain("Alice InPerson");
-    expect(html).toContain("Bob Remote");
-  });
-
-  test("shows all attendees when filter is all", () => {
-    const listing = testListingWithCount({ attendee_count: 2 });
-    const attendees = [
-      testAttendee({ checked_in: true, id: 1, name: "Checked In User" }),
-      testAttendee({ checked_in: false, id: 2, name: "Not Checked In User" }),
-    ];
-    const html = renderListingDetail({
-      activeFilter: "all",
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("Checked In User");
-    expect(html).toContain("Not Checked In User");
-  });
-
-  test("includes return_filter hidden field in checkin form", () => {
-    const listing = testListingWithCount({ attendee_count: 1 });
-    const attendees = [testAttendee({ checked_in: true })];
-    const html = renderListingDetail({
-      activeFilter: "in",
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain('name="return_filter"');
-    expect(html).toContain('value="in"');
-  });
-});
-
-describe("adminListingPage total revenue", () => {
-  test("shows total revenue for paid listings", () => {
-    const listing = testListingWithCount({
-      attendee_count: 2,
-      unit_price: 1000,
-    });
-    const attendees = [
-      testAttendee({ payment_id: "pi_test_1", price_paid: "1000" }),
-      testAttendee({ id: 2, payment_id: "pi_test_2", price_paid: "2000" }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("Total Revenue");
-    expect(html).toContain("£30");
-  });
-
-  test("does not show total revenue for free listings", () => {
-    const listing = testListingWithCount({ attendee_count: 1, unit_price: 0 });
-    const attendees = [testAttendee()];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).not.toContain("Total Revenue");
-  });
-
-  test("shows zero revenue for paid listing with no attendees", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      unit_price: 1000,
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Total Revenue");
-    expect(html).toContain("£0");
-  });
-});
-
-describe("adminListingPage income & ledger breakdown", () => {
-  const listing = testListingWithCount({ id: 7 });
-
-  test("omits the section entirely when no breakdown is supplied", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("Income &amp; ledger");
-  });
-
-  test("renders the five figures, the explanatory line and the ledger link", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-      revenueBreakdown: {
-        externalCosts: 0,
-        externalIncome: 0,
-        grossSales: 10000,
-        manualAdjustments: -1000,
-        netBalance: 7000,
-        recognisedIncome: 9000,
-        refunds: 2000,
-      },
-    });
-    expect(html).toContain("Income &amp; ledger");
-    // Gross sales credited (+), manual write-down (−), then the two subtotals.
-    expect(html).toContain("Gross ticket sales");
-    expect(html).toContain("+£100");
-    expect(html).toContain("Manual adjustments");
-    expect(html).toContain("−£10");
-    expect(html).toContain("Recognised income");
-    expect(html).toContain("£90");
-    expect(html).toContain("Refunds");
-    expect(html).toContain("−£20");
-    expect(html).toContain("Net balance in ledger");
-    expect(html).toContain("£70");
-    // The plain-English reconciliation note and the button to the filtered
-    // ledger, preselected to this listing (no arrow glyph, button-styled).
-    expect(html).toContain("refund-agnostic");
-    expect(html).toContain('href="/admin/ledger?listing=7"');
-    expect(html).toContain("View full ledger");
-    expect(html).not.toContain("View full ledger →");
-  });
-
-  test("makes a refund-driven divergence between income and net balance visible", () => {
-    // Recognised income (£90) and the net ledger balance (£70) legitimately
-    // differ after a refund; both must render so the reconciliation is shown.
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-      revenueBreakdown: {
-        externalCosts: 0,
-        externalIncome: 0,
-        grossSales: 9000,
-        manualAdjustments: 0,
-        netBalance: 7000,
-        recognisedIncome: 9000,
-        refunds: 2000,
-      },
-    });
-    const recognisedIdx = html.indexOf("Recognised income");
-    const netIdx = html.indexOf("Net balance in ledger");
-    expect(recognisedIdx).toBeGreaterThan(-1);
-    expect(netIdx).toBeGreaterThan(netIdx === -1 ? 0 : recognisedIdx);
-    expect(html).toContain("£90");
-    expect(html).toContain("£70");
-    // The two figures differ exactly by the refunds line.
-    expect(html).toContain("−£20");
-  });
-
-  test("omits the manual-adjustments row when there are none", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-      revenueBreakdown: {
-        externalCosts: 0,
-        externalIncome: 0,
-        grossSales: 5000,
-        manualAdjustments: 0,
-        netBalance: 5000,
-        recognisedIncome: 5000,
-        refunds: 0,
-      },
-    });
-    expect(html).toContain("Income &amp; ledger");
-    expect(html).not.toContain("Manual adjustments");
-    // With no refunds either, recognised income and net balance coincide at £50.
-    expect(html).toContain("Recognised income");
-    expect(html).toContain("Net balance in ledger");
-    expect(html).toContain("£50");
-  });
-
-  test("shows a signed positive manual write-up", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-      revenueBreakdown: {
-        externalCosts: 0,
-        externalIncome: 0,
-        grossSales: 4000,
-        manualAdjustments: 1500,
-        netBalance: 5500,
-        recognisedIncome: 5500,
-        refunds: 0,
-      },
-    });
-    expect(html).toContain("Manual adjustments");
-    expect(html).toContain("+£15");
-  });
-
-  test("shows outside income and listing-specific costs when present", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-      revenueBreakdown: {
-        externalCosts: 300,
-        externalIncome: 1200,
-        grossSales: 4000,
-        manualAdjustments: 0,
-        netBalance: 4900,
-        recognisedIncome: 5200,
-        refunds: 0,
-      },
-    });
-    expect(html).toContain("Income received outside checkout");
-    expect(html).toContain("+£12");
-    expect(html).toContain("Costs paid outside checkout");
-    expect(html).toContain("−£3");
-  });
-
-  test("shows a listing ledger add-entry action when the account exists", () => {
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      ledger: {
-        account: account("revenue", 7),
-        lines: [],
-        names: {
-          ...emptyLedgerNames(),
-          listings: new Map([[7, listing.name]]),
-        },
-      },
-      listing,
-    });
-    expect(html).toContain('<section id="ledger">');
-    expect(html).toContain("Add entry");
-    expect(html).toContain(
-      'href="/admin/ledger/revenue/7/add?return_url=%2Fadmin%2Flisting%2F7"',
-    );
-  });
-});
-
-describe("adminListingPage optional fields", () => {
-  test("hides thank you URL row when no thank_you_url", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      thank_you_url: "",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("Thank You URL");
-  });
-
-  test("shows webhook URL in copyable input when present", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      webhook_url: "https://hooks.example.com/notify",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Webhook URL");
-    expect(html).toContain('value="https://hooks.example.com/notify"');
-    expect(html).toContain("readonly");
-  });
-});
-
-describe("nearCapacity", () => {
-  test("returns true when at 90% capacity", () => {
-    const listing = testListingWithCount({
-      attendee_count: 90,
-      max_attendees: 100,
-    });
-    expect(nearCapacity(listing)).toBe(true);
-  });
-
-  test("returns true when over 90% capacity", () => {
-    const listing = testListingWithCount({
-      attendee_count: 95,
-      max_attendees: 100,
-    });
-    expect(nearCapacity(listing)).toBe(true);
-  });
-
-  test("returns false when under 90% capacity", () => {
-    const listing = testListingWithCount({
-      attendee_count: 89,
-      max_attendees: 100,
-    });
-    expect(nearCapacity(listing)).toBe(false);
-  });
-
-  test("returns true when fully sold out", () => {
-    const listing = testListingWithCount({
-      attendee_count: 100,
-      max_attendees: 100,
-    });
-    expect(nearCapacity(listing)).toBe(true);
-  });
-});
-
-describe("overviewStatsFromDbStats", () => {
-  const dbStats = {
-    completeQuantitySum: 5,
-    incompleteQuantity: 2,
-    incompleteSales: 300,
-    rowsCheckedIn: 1,
-    rowsTotal: 3,
-    ticketsCheckedIn: 2,
-    ticketsTotal: 4,
-  };
-
-  test("subtracts incomplete count and unpaid sales for a paid listing", () => {
-    const view = overviewStatsFromDbStats(dbStats, 7, 2100, true);
-    expect(view.adjustedCount).toBe(5); // 7 booked − 2 incomplete
-    expect(view.completeQuantitySum).toBe(5);
-    expect(view.completeRevenue).toBe(1800); // 2100 gross − 300 unpaid
-    expect(view.checkedInStats).toEqual({
-      hasMultiQuantity: true, // ticketsTotal 4 ≠ rowsTotal 3
-      rowsCheckedIn: 1,
-      rowsTotal: 3,
-      ticketsCheckedIn: 2,
-      ticketsTotal: 4,
-    });
-  });
-
-  test("reports zero revenue for a free listing and flat multi-quantity", () => {
-    const flat = { ...dbStats, rowsTotal: 4, ticketsTotal: 4 };
-    const view = overviewStatsFromDbStats(flat, 6, 999, false);
-    expect(view.completeRevenue).toBe(0);
-    expect(view.checkedInStats.hasMultiQuantity).toBe(false);
-  });
-});
-
-describe("isIncompletePayment", () => {
-  test("returns true for paid listing attendee with no payment_id and price > 0", () => {
-    const attendee = testAttendee({ payment_id: "", price_paid: "1000" });
-    expect(isIncompletePayment(attendee, true)).toBe(true);
-  });
-
-  test("returns false for free listing", () => {
-    const attendee = testAttendee({ payment_id: "", price_paid: "0" });
-    expect(isIncompletePayment(attendee, false)).toBe(false);
-  });
-
-  test("returns false for admin-added attendee on paid listing (price_paid=0)", () => {
-    const attendee = testAttendee({ payment_id: "", price_paid: "0" });
-    expect(isIncompletePayment(attendee, true)).toBe(false);
-  });
-
-  test("returns false for completed payment attendee", () => {
-    const attendee = testAttendee({
-      payment_id: "pi_test_123",
-      price_paid: "1000",
-    });
-    expect(isIncompletePayment(attendee, true)).toBe(false);
-  });
-});
-
-describe("completePaymentAttendees", () => {
-  test("drops unresolved-payment rows on a paid listing", () => {
-    const listing = testListingWithCount({ unit_price: 1000 });
-    const paid = testAttendee({
-      id: 1,
-      payment_id: "pi_ok",
-      price_paid: "1000",
-    });
-    const failed = testAttendee({ id: 2, payment_id: "", price_paid: "1000" });
-    expect(completePaymentAttendees(listing, [paid, failed])).toEqual([paid]);
-  });
-
-  test("keeps every row on a free listing", () => {
-    const listing = testListingWithCount({ unit_price: 0 });
-    const a = testAttendee({ id: 1, payment_id: "", price_paid: "0" });
-    const b = testAttendee({ id: 2, payment_id: "", price_paid: "1000" });
-    expect(completePaymentAttendees(listing, [a, b])).toEqual([a, b]);
-  });
-});
-
-describe("adminListingPage failed payments", () => {
-  test("shows Failed Payments section when incomplete attendees exist", () => {
-    const listing = testListingWithCount({
-      attendee_count: 3,
-      unit_price: 1000,
-    });
-    const attendees = [
-      testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-      testAttendee({ id: 2, payment_id: "", price_paid: "1000" }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("Failed Payments");
-    expect(html).toContain("1 attendee(s) with unresolved payments");
-    expect(html).toContain("/delete-incomplete");
-  });
-
-  test("hides Failed Payments section when no incomplete attendees", () => {
-    const listing = testListingWithCount({
-      attendee_count: 1,
-      unit_price: 1000,
-    });
-    const attendees = [
-      testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).not.toContain("Failed Payments");
-  });
-
-  test("hides Failed Payments section for free listings", () => {
-    const listing = testListingWithCount({ attendee_count: 1, unit_price: 0 });
-    const attendees = [testAttendee({ id: 1, price_paid: "0" })];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).not.toContain("Failed Payments");
-  });
-
-  test("excludes incomplete attendees from attendee count", () => {
-    const listing = testListingWithCount({
-      attendee_count: 3,
-      max_attendees: 100,
-      unit_price: 1000,
-    });
-    const attendees = [
-      testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-      testAttendee({ id: 2, payment_id: "", price_paid: "1000" }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    // adjusted count: 3 - 1 (incomplete qty) = 2
-    expect(html).toContain("2 / 100");
-  });
-
-  test("excludes incomplete attendees from checked-in count", () => {
-    const listing = testListingWithCount({
-      attendee_count: 2,
-      unit_price: 1000,
-    });
-    const attendees = [
-      testAttendee({
-        checked_in: true,
-        id: 1,
-        payment_id: "pi_ok",
-        price_paid: "1000",
-      }),
-      testAttendee({
-        checked_in: true,
-        id: 2,
-        payment_id: "",
-        price_paid: "1000",
-      }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    // Only complete attendees count: 1 checked in / 1 total
-    expect(html).toContain("1 / 1");
-  });
-
-  test("excludes incomplete attendees from revenue", () => {
-    const listing = testListingWithCount({
-      attendee_count: 2,
-      unit_price: 1000,
-    });
-    const attendees = [
-      testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-      testAttendee({ id: 2, payment_id: "", price_paid: "2000" }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    expect(html).toContain("£10");
-    expect(html).not.toContain("£30");
-  });
-
-  test("failed payments table has delete button but no check-in or refund", () => {
-    const listing = testListingWithCount({
-      attendee_count: 1,
-      unit_price: 1000,
-    });
-    const attendees = [
-      testAttendee({
-        id: 1,
-        name: "Jane Stuck",
-        payment_id: "",
-        price_paid: "1000",
-      }),
-    ];
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees,
-      listing,
-    });
-    const failedSection =
-      html.split("Failed Payments")[1]?.split("Add Attendee")[0] ?? "";
-    expect(failedSection).toContain("Jane Stuck");
-    expect(failedSection).toContain("Delete");
-    expect(failedSection).toContain("/delete-incomplete");
-    expect(failedSection).not.toContain("Check in");
-    expect(failedSection).not.toContain("Check out");
-    expect(failedSection).not.toContain("Refund");
-    expect(failedSection).not.toContain("Re-send Webhook");
-  });
-});
-
-describe("adminListingPage listing date and location", () => {
-  test("shows Listing Date row when listing has a date", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      date: "2026-06-15T14:00:00.000Z",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Listing Date");
-    expect(html).toContain("Monday 15 June 2026 at 15:00 GMT+1");
-  });
-
-  test("does not show Listing Date row when date is empty", () => {
-    const listing = testListingWithCount({ attendee_count: 0, date: "" });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("Listing Date");
-  });
-
-  test("shows Location row when listing has a location", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      location: "Village Hall",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("<th>Location</th>");
-    expect(html).toContain("Village Hall");
-  });
-
-  test("does not show Location row when location is empty", () => {
-    const listing = testListingWithCount({ attendee_count: 0, location: "" });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("<th>Location</th>");
-  });
-
-  test("shows both Listing Date and Location when both are set", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      date: "2026-06-15T14:00:00.000Z",
-      location: "Town Centre",
-    });
-    const html = renderListingDetail({
-      allowedDomain: "localhost",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Listing Date");
-    expect(html).toContain("Town Centre");
-  });
-});
-
-describe("adminListingPage edit form pre-fills date and location", () => {
-  test("empty date shows no pre-filled value in edit form", () => {
-    const listing = testListingWithCount({ attendee_count: 0, date: "" });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    // The date field should render split date and time inputs
-    expect(html).toContain('name="date_date"');
-    expect(html).toContain('name="date_time"');
-  });
-
-  test("non-empty date shows formatted split values in edit form", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      date: "2026-06-15T14:00:00.000Z",
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    // Should contain split date and time values converted to Europe/London (BST = UTC+1)
-    expect(html).toContain('value="2026-06-15"');
-    expect(html).toContain('value="15:00"');
-  });
-
-  test("pre-fills location in edit form", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      location: "Village Hall",
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain('value="Village Hall"');
-  });
-});
-
-describe("adminListingEditPage max_price field", () => {
-  test("renders max_price field with value when set", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      can_pay_more: true,
-      max_price: 50000,
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain('name="max_price"');
-    expect(html).toContain('value="500.00"');
-  });
-
-  test("renders max_price field with 0.00 when zero", () => {
-    const listing = testListingWithCount({
-      attendee_count: 0,
-      can_pay_more: true,
-      max_price: 0,
-    });
-    const html = String(
-      ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-    );
-    expect(html).toContain('name="max_price"');
-    expect(html).toContain('value="0.00"');
-  });
-});
-
-describe("datetime validation via getListingFields() date field", () => {
-  const dateField = getListingFields().find((f) => f.name === "date")!;
-
-  test("accepts valid datetime value", () => {
-    const result = dateField.validate?.("2026-06-15T14:00");
-    expect(result).toBeNull();
-  });
-
-  test("rejects invalid datetime value", () => {
-    const result = dateField.validate?.("not-a-date");
-    expect(result).toBe("Please enter a valid date and time");
-  });
-});
-
-describeWithEnv(
-  "listing images",
-  { env: { STORAGE_ZONE_KEY: "testkey", STORAGE_ZONE_NAME: "testzone" } },
-  () => {
-    describe("adminListingPage image section", () => {
-      test("does not show image upload on detail page", () => {
-        const listing = testListingWithCount({ image_url: "" });
-        const html = renderListingDetail({
-          allowedDomain: "localhost",
-          attendees: [],
-          listing,
-        });
-        expect(html).not.toContain('type="file"');
-        expect(html).not.toContain('name="image"');
-      });
-    });
-
-    describe("listing form image ownership", () => {
-      test("does not show image controls on the edit form", () => {
-        runWithStorageConfig(TEST_STORAGE_ZONE, () => {
-          const listing = testListingWithCount({ image_url: "current.jpg" });
-          const html = String(
-            ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-          );
-          expect(html).not.toContain('name="image"');
-          expect(html).not.toContain("Remove Image");
-          expect(html).not.toContain("/image/delete");
-          expect(html).not.toContain("listing-image-full");
-        });
-      });
-
-      test("does not show image controls on duplicate or create forms", () => {
-        runWithStorageConfig(TEST_STORAGE_ZONE, () => {
-          const listing = testListingWithCount({ image_url: "current.jpg" });
-          expect(
-            adminDuplicateListingPage(listing, [], TEST_SESSION),
-          ).not.toContain('name="image"');
-          expect(adminListingNewPage([], TEST_SESSION)).not.toContain(
-            'name="image"',
-          );
-        });
-      });
-
-      test("keeps image controls absent when storage is disabled", () => {
-        withStorageDisabled(() => {
-          const listing = testListingWithCount({ image_url: "current.jpg" });
-          expect(
-            String(
-              ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-            ),
-          ).not.toContain('name="image"');
-          expect(
-            adminDuplicateListingPage(listing, [], TEST_SESSION),
-          ).not.toContain('name="image"');
-          expect(adminListingNewPage([], TEST_SESSION)).not.toContain(
-            'name="image"',
-          );
-        });
-      });
-    });
-
-    describe("assign_built_site field", () => {
-      test("shows assign built site field when CAN_BUILD_SITES is true", () => {
-        withBuilder(() => {
-          const html = adminListingNewPage([], TEST_SESSION);
-          expect(html).toContain("assign_built_site");
-          expect(html).toContain("Assign a site on booking");
-        });
-      });
-
-      test("hides assign built site field when CAN_BUILD_SITES is not set", () => {
-        withoutBuilder(() => {
-          const html = adminListingNewPage([], TEST_SESSION);
-          expect(html).not.toContain("assign_built_site");
-        });
-      });
-
-      test("shows on edit page when CAN_BUILD_SITES is true", () => {
-        withBuilder(() => {
-          const listing = testListingWithCount({ assign_built_site: true });
-          const html = String(
-            ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-          );
-          expect(html).toContain("assign_built_site");
-          expect(html).toContain("checked");
-        });
-      });
-
-      test("shows on duplicate page when CAN_BUILD_SITES is true", () => {
-        withBuilder(() => {
-          const listing = testListingWithCount({ assign_built_site: true });
-          const html = adminDuplicateListingPage(listing, [], TEST_SESSION);
-          expect(html).toContain("assign_built_site");
-        });
-      });
-    });
-
-    describe("months_per_unit and initial_site_months fields", () => {
-      test("shows months_per_unit and initial_site_months when CAN_BUILD_SITES is true", () => {
-        withBuilder(() => {
-          const html = adminListingNewPage([], TEST_SESSION);
-          expect(html).toContain("months_per_unit");
-          expect(html).toContain("Months Per Unit");
-          expect(html).toContain("initial_site_months");
-          expect(html).toContain("Initial Site Months");
-        });
-      });
-
-      test("hides months_per_unit and initial_site_months when CAN_BUILD_SITES is not set", () => {
-        withoutBuilder(() => {
-          const html = adminListingNewPage([], TEST_SESSION);
-          expect(html).not.toContain("months_per_unit");
-          expect(html).not.toContain("Months Per Unit");
-          expect(html).not.toContain("initial_site_months");
-          expect(html).not.toContain("Initial Site Months");
-        });
-      });
-
-      test("shows on edit page when CAN_BUILD_SITES is true", () => {
-        withBuilder(() => {
-          const listing = testListingWithCount({
-            initial_site_months: 6,
-            months_per_unit: 3,
-          });
-          const html = String(
-            ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-          );
-          expect(html).toContain("months_per_unit");
-          expect(html).toContain("initial_site_months");
-        });
-      });
-
-      test("hides on edit page when CAN_BUILD_SITES is not set", () => {
-        withoutBuilder(() => {
-          const listing = testListingWithCount({
-            initial_site_months: 6,
-            months_per_unit: 3,
-          });
-          const html = String(
-            ListingEditPanel({ groups: [], listing, session: TEST_SESSION }),
-          );
-          expect(html).not.toContain("months_per_unit");
-          expect(html).not.toContain("initial_site_months");
-        });
-      });
-
-      test("shows on duplicate page when CAN_BUILD_SITES is true", () => {
-        withBuilder(() => {
-          const listing = testListingWithCount({
-            initial_site_months: 6,
-            months_per_unit: 3,
-          });
-          const html = adminDuplicateListingPage(listing, [], TEST_SESSION);
-          expect(html).toContain("months_per_unit");
-          expect(html).toContain("initial_site_months");
-        });
-      });
-
-      test("hides on duplicate page when CAN_BUILD_SITES is not set", () => {
-        withoutBuilder(() => {
-          const listing = testListingWithCount({
-            initial_site_months: 6,
-            months_per_unit: 3,
-          });
-          const html = adminDuplicateListingPage(listing, [], TEST_SESSION);
-          expect(html).not.toContain("months_per_unit");
-          expect(html).not.toContain("initial_site_months");
-        });
-      });
-    });
-  },
-);
-
-describe("adminListingPage Renewal tag", () => {
-  test("renders Renewal tag for tier listings with months_per_unit > 0", () => {
-    const listing = testListingWithCount({ months_per_unit: 3 });
-    const html = renderListingDetail({
-      allowedDomain: "",
-      attendees: [],
-      listing,
-    });
-    expect(html).toContain("Renewal");
-  });
-
-  test("does not render Renewal tag for listings with months_per_unit = 0", () => {
-    const listing = testListingWithCount({ months_per_unit: 0 });
-    const html = renderListingDetail({
-      allowedDomain: "",
-      attendees: [],
-      listing,
-    });
-    expect(html).not.toContain("Renewal");
-  });
-});
